# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/conversations:
    get:
      operationId: listConversations
      tags: [messages]
      summary: List chat conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: startConversation
      tags: [messages]
      summary: Start a conversation by username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationInput"
      responses:
        "201":
          description: Conversation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationStarted"
  /messages/conversations/{id}:
    get:
      operationId: getConversation
      tags: [messages]
      summary: Get conversation messages
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: closeConversation
      tags: [messages]
      summary: Close a conversation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Closed
  /messages/conversations/{conversationId}/gig-invites:
    get:
      operationId: listConversationGigInvites
      tags: [messages]
      summary: List pending gig invites in this conversation (artist only)
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pending gig invites from this venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GigInviteList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not an artist or not a participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /messages/conversations/{conversationId}/gig-invites/{outreachId}/respond:
    post:
      operationId: respondToGigInvite
      tags: [messages]
      summary: Accept or decline a gig invite (artist only)
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: integer
        - in: path
          name: outreachId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RespondToGigInviteInput"
      responses:
        "200":
          description: Updated outreach entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventOutreachEntry"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /messages/users/search:
    get:
      operationId: searchChatUsers
      tags: [messages]
      summary: Search users to start a chat
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatUserSearchList"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    EventOutreachEntry:
      type: object
      required: [id, artistId, displayName, status]
      properties:
        id:
          type: integer
        artistId:
          type: integer
        displayName:
          type: string
        username:
          type: string
        genres:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [contacted, pending, confirmed, declined]
        notes:
          type: ["string", "null"]
        eventId:
          type: integer
        eventTitle:
          type: string
        eventDate:
          type: string
          format: date-time
...
    GigInviteList:
      type: object
      required: [invites]
      properties:
        invites:
          type: array
          items:
            $ref: "#/components/schemas/GigInvite"
    GigInvite:
      type: object
      required: [outreachId, eventId, eventTitle, eventDate, venueId, venueName, status]
      properties:
        outreachId:
          type: integer
        eventId:
          type: integer
        eventTitle:
          type: string
        eventDate:
          type: string
          format: date-time
        venueId:
          type: integer
        venueName:
          type: string
        status:
          type: string
          enum: [pending]
    RespondToGigInviteInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, decline]
...
    ChatUser:
      type: object
      required: [id, username, displayName, role]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        profileId:
          type: ["integer", "null"]
          description: Public profile id (venue id or artist id)
    ChatUserSearchList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/ChatUser"
    Message:
      type: object
      required: [id, senderUserId, createdAt]
      properties:
        id:
          type: integer
        senderUserId:
          type: integer
        body:
          type: ["string", "null"]
        attachmentUrl:
          type: ["string", "null"]
        attachmentType:
          type: ["string", "null"]
        attachmentName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ConversationSummary:
      type: object
      required: [id, otherUser, updatedAt]
      properties:
        id:
          type: integer
        otherUser:
          $ref: "#/components/schemas/ChatUser"
        lastMessage:
          type: ["object", "null"]
          properties:
            body:
              type: ["string", "null"]
            attachmentType:
              type: ["string", "null"]
            createdAt:
              type: string
              format: date-time
        updatedAt:
          type: string
          format: date-time
    ConversationList:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/ConversationSummary"
    ConversationDetail:
      type: object
      required: [id, messages]
      properties:
        id:
          type: integer
        otherUser:
          oneOf:
            - $ref: "#/components/schemas/ChatUser"
            - type: "null"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    ConversationStarted:
      type: object
      required: [id, otherUser]
      properties:
        id:
          type: integer
        otherUser:
          $ref: "#/components/schemas/ChatUser"
    StartConversationInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
    SendMessageInput:
      type: object
      properties:
        body:
          type: string
        attachmentUrl:
          type: string
        attachmentType:
          type: string
          enum: [image, file]
        attachmentName:
          type: string
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid