# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatUser } from './chatUser';
import type { Message } from './message';

export interface ConversationDetail {
  id: number;
  otherUser?: ChatUser | null;
  messages: Message[];
}