# lib/api-zod/src/generated/types/conversationList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationSummary } from './conversationSummary';

export interface ConversationList {
  conversations: ConversationSummary[];
}