# lib/api-zod/src/generated/types/eventOutreachEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventOutreachEntryStatus } from './eventOutreachEntryStatus';

export interface EventOutreachEntry {
  id: number;
  artistId: number;
  displayName: string;
  username?: string;
  genres?: string[];
  status: EventOutreachEntryStatus;
  /** @nullable */
  notes?: string | null;
  eventId?: number;
  eventTitle?: string;
  eventDate?: Date;
}